+++ craft_ai/pandas/client.py
@@ -167,7 +167,7 @@ class Client(VanillaClient):
         )
 
     @staticmethod
-    def decide_from_contexts_df(tree, contexts_df):
+    def check_decision_context_df(contexts_df):
         if isinstance(contexts_df, pd.DataFrame):
             if contexts_df.empty:
                 raise CraftAiBadRequestError(
@@ -184,6 +184,10 @@ class Client(VanillaClient):
                 )
         else:
             raise CraftAiBadRequestError("Invalid data given, it is not a DataFrame.")
+
+    @staticmethod
+    def decide_from_contexts_df(tree, contexts_df):
+        Client.check_decision_context_df(contexts_df)
         return Interpreter.decide_from_contexts_df(tree, contexts_df)
 
     def get_agent_decision_tree(
@@ -194,6 +198,41 @@ class Client(VanillaClient):
             timestamp = timestamp.value // 10 ** 9
 
         return super(Client, self).get_agent_decision_tree(agent_id, timestamp, version)
+
+    def pandas_agent_boosting_decide_from_row(
+        self, generator_id, from_ts, to_ts, params
+    ):
+        context = {
+            feature_name: format_input(value)
+            for feature_name, value in zip(
+                params["feature_names"], params["context_ops"][1:]
+            )
+            if is_valid_property_value(feature_name, value)
+        }
+        try:
+            decision = super(Client, self).get_agent_boosting_decision(
+                generator_id, from_ts, to_ts, context
+            )
+            return {"output_predicted_value": decision["output"]["predicted_value"]}
+
+        except CraftAiNullDecisionError as e:
+            return {"error": e.message}
+
+    def decide_boosting_from_contexts_df(
+        self, generator_id, from_ts, to_ts, contexts_df
+    ):
+        Client.check_decision_context_df(contexts_df)
+        df = contexts_df.copy(deep=True)
+        predictions_iter = (
+            self.pandas_agent_boosting_decide_from_row(
+                generator_id,
+                from_ts,
+                to_ts,
+                {"context_ops": row, "feature_names": df.columns.values},
+            )
+            for row in df.itertuples(name=None)
+        )
+        return pd.DataFrame(predictions_iter, index=df.index)
 
     def get_generator_decision_tree(
         self, generator_id, timestamp=None, version=DEFAULT_DECISION_TREE_VERSION
@@ -206,7 +245,9 @@ class Client(VanillaClient):
             generator_id, timestamp, version
         )
 
-    def pandas_generator_decide_from_row(self, generator_id, from_ts, to_ts, params):
+    def pandas_generator_boosting_decide_from_row(
+        self, generator_id, from_ts, to_ts, params
+    ):
         context = {
             feature_name: format_input(value)
             for feature_name, value in zip(
@@ -226,27 +267,11 @@ class Client(VanillaClient):
     def decide_generator_boosting_from_contexts_df(
         self, generator_id, from_ts, to_ts, contexts_df
     ):
-        if isinstance(contexts_df, pd.DataFrame):
-            if contexts_df.empty:
-                raise CraftAiBadRequestError(
-                    "Invalid dataframe given, dataframe is empty."
-                )
-            if not isinstance(contexts_df.index, pd.DatetimeIndex):
-                raise CraftAiBadRequestError(
-                    "Invalid dataframe given, it is not time indexed."
-                )
-            if contexts_df.index.tz is None:
-                raise CraftAiBadRequestError(
-                    """tz-naive DatetimeIndex are not supported,
-                                     it must be tz-aware."""
-                )
-        else:
-            raise CraftAiBadRequestError("Invalid data given, it is not a DataFrame.")
-
+        Client.check_decision_context_df(contexts_df)
         df = contexts_df.copy(deep=True)
 
         predictions_iter = (
-            self.pandas_generator_decide_from_row(
+            self.pandas_generator_boosting_decide_from_row(
                 generator_id,
                 from_ts,
                 to_ts,